+++ src/main/java/db/data/types/DataType.java
@@ -1,4 +1,4 @@
-package db.data;
+package db.data.types;
 
 import java.io.DataInput;
 import java.io.IOException;
