+++ src/main/java/db/search/FilterTerm.java
@@ -1,6 +1,6 @@
 package db.search;
 
-import db.data.DataPositionList;
+import db.data.types.DataPositionList;
 
 public interface FilterTerm {
     DataPositionList execute() throws SearchException;
