+++ src/test/java/db/sTreeIndex/IndexTreeMessyTest.java
@@ -20,14 +20,14 @@ import com.dant.utils.MemUsage;
 import com.dant.utils.Timer;
 import com.dant.utils.Utils;
 
-import db.data.ByteType;
-import db.data.DateType;
-import db.data.DoubleType;
-import db.data.FloatType;
-import db.data.IntegerArrayList;
-import db.data.DataPositionList;
-import db.data.StringType;
-import db.parsers.Parser;
+import db.data.types.ByteType;
+import db.data.types.DateType;
+import db.data.types.DoubleType;
+import db.data.types.FloatType;
+import db.data.types.IntegerArrayList;
+import db.data.types.DataPositionList;
+import db.data.types.StringType;
+import db.data.load.Parser;
 import db.structure.Column;
 import db.structure.Database;
 import db.structure.Table;
