+++ src/test/java/db/sTreeIndex/TestIndexTreeMT1.java
@@ -3,7 +3,6 @@ package db.sTreeIndex;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
@@ -19,14 +18,13 @@ import com.dant.utils.MemUsage;
 import com.dant.utils.Timer;
 import com.dant.utils.Utils;
 
-import db.data.ByteType;
-import db.data.DateType;
-import db.data.DoubleType;
-import db.data.FloatType;
-import db.data.IntegerArrayList;
-import db.data.DataPositionList;
-import db.data.StringType;
-import db.parsers.CsvParser;
+import db.data.types.ByteType;
+import db.data.types.DateType;
+import db.data.types.DoubleType;
+import db.data.types.FloatType;
+import db.data.types.IntegerArrayList;
+import db.data.types.StringType;
+import db.data.load.CsvParser;
 import db.structure.Column;
 import db.structure.Table;
 import db.structure.indexTree.IndexTreeCeption;
